# docs/swagger.yaml
basePath: /api/v1
definitions:
  common.ResponseHTTP:
    properties:
      data: {}
      details:
        type: string
      success:
        type: boolean
    type: object
  common.ResponsePagination:
    properties:
      data: {}
      details:
        type: string
      page:
        type: integer
      pages:
        type: integer
      size:
        type: integer
      success:
        type: boolean
      total:
        type: integer
    type: object
  controllers.AppEndpointsMeta:
    additionalProperties:
      items:
        type: string
      type: array
    type: object
...
  controllers.LoginPost:
    properties:
      email:
        maxLength: 32
        minLength: 6
        type: string
      grant_type:
        type: string
      password:
        type: string
      token:
        type: string
    required:
    - grant_type
    type: object
...
  controllers.TokenResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
...
  models.AppGet:
    description: AppGet type information
    properties:
      active:
        type: boolean
      description:
        type: string
      id:
        type: integer
      name:
        type: string
      roles:
        items:
          $ref: '#/definitions/models.Role'
        type: array
      uuid:
        type: string
    type: object
  models.AppPost:
    description: AppPost type information
    properties:
      description:
        type: string
      name:
        type: string
    type: object
...
info:
  contact: {}
  description: This is blue-admin API OPENAPI Documentation.
  termsOfService: http://swagger.io/terms/
  title: Swagger blue-admin API
  version: "0.1"
paths:
  /app:
    get:
      consumes:
      - application/json
      description: Get Apps
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: page size
        in: query
        name: size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponsePagination'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.AppGet'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      - Refresh: []
      summary: Get Apps
      tags:
      - Apps
    post:
      consumes:
      - application/json
      description: Add App
      parameters:
      - description: Add App
        in: body
        name: app
        required: true
        schema:
          $ref: '#/definitions/models.AppPost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/models.AppPost'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Add a new App
      tags:
      - Apps
  /app/{app_id}:
    delete:
      consumes:
      - application/json
      description: Remove app by ID
      parameters:
      - description: App ID
        in: path
        name: app_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Remove App by ID
      tags:
      - Apps
    get:
      consumes:
      - application/json
      description: Get app by ID
      parameters:
      - description: App ID
        in: path
        name: app_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/models.AppGet'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Get App by ID
      tags:
      - Apps
    patch:
      consumes:
      - application/json
      description: Patch App
      parameters:
      - description: Patch App
        in: body
        name: app
        required: true
        schema:
          $ref: '#/definitions/models.AppPost'
      - description: App ID
        in: path
        name: app_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/models.AppPost'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Patch App
      tags:
      - Apps
  /checklogin:
    get:
      consumes:
      - application/json
      description: CheckLogin
      parameters:
      - description: Login
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/controllers.LoginPost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/controllers.TokenResponse'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      summary: Auth
      tags:
      - Authentication
  /dashboard:
    get:
      consumes:
      - application/json
      description: Get App summary by App ID
      parameters:
      - description: App ID
        in: query
        name: app_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/controllers.AppMeta'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Get Roles Grouped By APP ID
      tags:
      - Dashboard Meta
  /dashboardends:
    get:
      consumes:
      - application/json
      description: Get App Endpoints summary
      parameters:
      - description: App ID
        in: query
        name: app_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/controllers.AppEndpointsMeta'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Get App Endpoints Grouped By Role
      tags:
      - Dashboard Meta
  /dashboardfeat:
    get:
      consumes:
      - application/json
      description: Get App Featuers summary
      parameters:
      - description: App ID
        in: query
        name: app_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/controllers.AppFeaturesMeta'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Get App Features Grouped By Role
      tags:
      - Dashboard Meta
  /dashboardpages:
    get:
      consumes:
      - application/json
      description: Get App Pages summary
      parameters:
      - description: App ID
        in: query
        name: app_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/controllers.AppFeaturesMeta'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Get App Pages
      tags:
      - Dashboard Meta
  /dashboardroles:
    get:
      consumes:
      - application/json
      description: Get App Roles summary
      parameters:
      - description: App ID
        in: query
        name: app_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/controllers.AppFeaturesMeta'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Get App Roles
      tags:
      - Dashboard Meta
  /dashboardrolespage:
    get:
      consumes:
      - application/json
      description: Get Roles by page by ID
      parameters:
      - description: App ID
        in: query
        name: app_id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/models.PageGet'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Get Page Roles for Secfic App by ID
      tags:
      - Dashboard Meta
  /droproles:
    get:
      consumes:
      - application/json
      description: Get RoleDropDown
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/controllers.RoleDropDown'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Get RoleDropDown
      tags:
      - Role
  /email:
    post:
      consumes:
      - application/json
      description: Sending Email
      parameters:
      - description: messages
        in: body
        name: User
        required: true
        schema:
          $ref: '#/definitions/messages.EmailMessage'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/messages.EmailMessage'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Send Email to
      tags:
      - Utilities
  /endpoint:
    get:
      consumes:
      - application/json
      description: Get Endpoints
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: page size
        in: query
        name: size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponsePagination'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.EndpointGet'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      - Refresh: []
      summary: Get Endpoints
      tags:
      - Endpoints
    post:
      consumes:
      - application/json
      description: Add Endpoint
      parameters:
      - description: Add Endpoint
        in: body
        name: endpoint
        required: true
        schema:
          $ref: '#/definitions/models.EndpointPost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/models.EndpointPost'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Add a new Endpoint
      tags:
      - Endpoints
  /endpoint/{endpoint_id}:
    delete:
      consumes:
      - application/json
      description: Remove endpoint by ID
      parameters:
      - description: Endpoint ID
        in: path
        name: endpoint_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Remove Endpoint by ID
      tags:
      - Endpoints
    get:
      consumes:
      - application/json
      description: Get endpoint by ID
      parameters:
      - description: Endpoint ID
        in: path
        name: endpoint_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/models.EndpointGet'
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Get Endpoint by ID
      tags:
      - Endpoints
    patch:
      consumes:
      - application/json
      description: Patch Endpoint
      parameters:
      - description: Patch Endpoint
        in: body
        name: endpoint
        required: true
        schema:
          $ref: '#/definitions/models.EndpointPost'
      - description: Endpoint ID
        in: path
        name: endpoint_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/models.EndpointPost'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Patch Endpoint
      tags:
      - Endpoints
...
  /endpointfeature/{endpoint_id}:
    delete:
      consumes:
      - application/json
      description: Delete Feature Endpoint
      parameters:
      - description: Feature ID
        in: path
        name: endpoint_id
        required: true
        type: integer
      - description: Endpoint ID
        in: query
        name: feature_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Delete Feature Endpoint
      tags:
      - Features
    patch:
      consumes:
      - application/json
      description: Add Feature to Endpoint
      parameters:
      - description: Endpoint ID
        in: path
        name: endpoint_id
        required: true
        type: integer
      - description: ' Feature ID'
        in: query
        name: feature_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Add Feature to Endpoint
      tags:
      - Features
  /feature:
    get:
      consumes:
      - application/json
      description: Get Features
      parameters:
      - description: page
        in: query
        name: page
        required: true
        type: integer
      - description: page size
        in: query
        name: size
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponsePagination'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/models.FeatureGet'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      - Refresh: []
      summary: Get Features
      tags:
      - Features
    post:
      consumes:
      - application/json
      description: Add Feature
      parameters:
      - description: Add Feature
        in: body
        name: feature
        required: true
        schema:
          $ref: '#/definitions/models.FeaturePost'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/models.FeaturePost'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Add a new Feature
      tags:
      - Features
...
  /featuredrop:
    get:
      consumes:
      - application/json
      description: Get FeatureDropDown
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/controllers.FeatureDropDown'
                  type: array
              type: object
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Get FeatureDropDown
      tags:
      - Feature
...
  /features/{feature_id}:
    put:
      consumes:
      - application/json
      description: Activate/Deactivate Feature
      parameters:
      - description: Feature ID
        in: path
        name: feature_id
        required: true
        type: integer
      - description: Active
        in: query
        name: active
        required: true
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/common.ResponseHTTP'
            - properties:
                data:
                  $ref: '#/definitions/models.FeaturePost'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/common.ResponseHTTP'
      security:
      - ApiKeyAuth: []
      summary: Activate/Deactivate Feature
      tags:
      - Feature
...
securityDefinitions:
  ApiKeyAuth:
    description: Description for what is this security definition being used
    in: header
    name: X-APP-TOKEN
    type: apiKey
  Refresh:
    description: Description for what is this security definition being used
    in: header
    name: X-REFRESH-TOKEN
    type: apiKey
swagger: "2.0"